docs: add usage.md (#280)
Adds examples of how to interact with nz-elevation from various
applications.

This currently duplicates a lot of content on the linz/imagery repo but
links and images have been updated specifically for linz/elevation.

We will at some point want to consolidate docs to reduce duplication but
in the interim it will be good to link directly from the Registry of
Open Data on AWS listing to these docs.

---------

Co-authored-by: Alice Fage <alice.fage@gmail.com>