ci: remove generate-name & exclude collection TDE-1058 (#282)
The logic to use $base (the name of the parameters.yaml file) to name
the argo workflow submitted no longer works.
This change will give each submission the name `publish-odr-file-copy-`
(so that can be distinguished from publihs-odr run by any other means)

In addition, the collection will be "synced" later on this workflow and
include any changed made in the PR that will not be in the /flat/
directory. Therefore we should exlude it here.


Related linz/imagery/ changes:
https://github.com/linz/imagery/pull/377
https://github.com/linz/imagery/pull/379